requirements.txt
addict==2.4.0
diffusers==0.15.1
gradio==4.29.0
huggingface_hub==0.23.0
matplotlib==3.6.0
numpy
onnxruntime==1.14.1
opencv_python==4.7.0.72
Pillow==10.3.0
pycocotools==2.0.6
PyYAML==6.0
requests==2.31.0
setuptools==65.5.0
supervision==0.20.0
termcolor
timm==0.9.16
torch==1.13.1
torchvision==0.14.1
transformers==4.40.1
yapf==0.40.2
nltk
fairscale
litellm

